feat: move document data to appstate
* feat: add status page showing if clicking and mouse pos
* feat: implement reactive string data for status page
* chore: remove printf statement

NOTE: Garbage was showing up in the status page, this may be due to how
I'm handling strings. I implemented a couple of `memset`s, so hopefully
that solves the issue. More work may need to be done - keep an eye out
for garbage in the status window from the strings.